Refactored Setting page (#61)
